Cast uint16_t ADC as uint32_t

+++ HipMounted_Pedometer/Core/Inc/main.h
@@ -60,7 +60,7 @@ extern volatile uint32_t stepCount;
 extern volatile uint32_t indexVal; // Global index variable
 extern volatile int distanceTravelled; // Distance travelled [m]
 
-extern uint32_t ADC_VAL[3];
+extern uint16_t ADC_VAL[3];
 extern float adjVal[2][3];
 
 extern UART_HandleTypeDef huart2;
